+++ gestor-pedidos-lanchonete/sistema-de-gestao-de-pedidos/src/main/java/com/lanchonete/repository/VendedorRepository.java
@@ -1,10 +1,21 @@
+/**
+ * VendedorRepository
+ * -------------------
+ * O que ela altera no programa:
+ * - Controla onde e como os vendedores são guardados.
+ * - Permite que o restante do sistema trabalhe com vendedores
+ *   sem se preocupar com armazenamento.
+ */
+
 package com.lanchonete.repository;
 
-import com.lanchonete.model.Vendedor;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.lanchonete.model.Vendedor;
+
 public class VendedorRepository implements IVendedorRepository {
+
     private final List<Vendedor> vendedores = new ArrayList<>();
 
     @Override
@@ -20,7 +31,7 @@ public class VendedorRepository implements IVendedorRepository {
     }
 
     @Override
-    public Vendedor buscarPorCodigo(double codigo) {
+    public Vendedor buscarPorCodigo(int codigo) {
         return vendedores.stream()
                 .filter(v -> v.getCodigo() == codigo)
                 .findFirst()
